test : 기본 토큰 생성 테스트 로직 작성 및 step 메소드들 리팩토링
